[REFACTOR] Merge Service services, clean up controller
