+++ Eventy/src/main/java/org/example/eventy/solutions/controllers/ServiceController.java
@@ -3,7 +3,6 @@ package org.example.eventy.solutions.controllers;
 import org.example.eventy.events.dtos.EventTypeDTO;
 import org.example.eventy.solutions.dtos.CategoryDTO;
 import org.example.eventy.solutions.dtos.services.*;
-import org.example.eventy.solutions.services.ServicesService;
 import org.example.eventy.solutions.dtos.SolutionCardDTO;
 import org.example.eventy.solutions.services.ServiceService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,35 +20,32 @@ import java.util.ArrayList;
 @RequestMapping("api/services")
 public class ServiceController {
 
-    private ServicesService servicesService;
-
-    public ServiceController() {
-        servicesService = new ServicesService();
-    }
+    @Autowired
+    private ServiceService serviceService;
 
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<CreatedServiceDTO> createService(@RequestBody CreateServiceDTO service) {
-        CreatedServiceDTO response = servicesService.createService(service);
+        CreatedServiceDTO response = serviceService.createService(service);
 
         return new ResponseEntity<CreatedServiceDTO>(response, HttpStatus.CREATED);
     }
 
     // NOTE: Add extra RequestParams as necessary if more filters are needed
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    /*@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Collection<GetServiceDTO>> getServices(@RequestParam (required = false) String name,
                                                             @RequestParam (required = false) CategoryDTO category,
                                                             @RequestParam (required = false) EventTypeDTO eventType,
                                                             @RequestParam (required = false) double minPrice,
                                                             @RequestParam (required = false) double maxPrice,
                                                             @RequestParam (required = false) boolean available) {
-        Collection<GetServiceDTO> services = servicesService.getServices(name, category, eventType, minPrice, maxPrice, available);
+        Collection<GetServiceDTO> services = serviceService.getServices(name, category, eventType, minPrice, maxPrice, available);
 
         return new ResponseEntity<>(services, HttpStatus.OK);
-    }
+    }*/
 
     @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<GetServiceDTO> getService(@PathVariable("id") long id) {
-        Optional<GetServiceDTO> service = servicesService.getService(id);
+        Optional<GetServiceDTO> service = serviceService.getService(id);
         if (!service.isPresent()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -59,7 +55,7 @@ public class ServiceController {
     @PutMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<UpdatedServiceDTO> updateService(@RequestBody UpdateServiceDTO service, @PathVariable("id") long id) {
         service.setId(id);
-        Optional<UpdatedServiceDTO> updatedService = servicesService.updateService(service);
+        Optional<UpdatedServiceDTO> updatedService = serviceService.updateService(service);
         if (!updatedService.isPresent()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
@@ -68,12 +64,9 @@ public class ServiceController {
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> deleteService(@PathVariable("id") long id) {
-        servicesService.deleteService(id);
+        serviceService.deleteService(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
-    @Autowired
-    private ServiceService serviceService;
 
     /* this returns SolutionCardDTOs, because there is NO CASE where:
        1) we need ALL services
