+++ Eventy/src/main/java/org/example/eventy/solutions/services/ServiceService.java
@@ -1,17 +1,143 @@
 package org.example.eventy.solutions.services;
 
+import org.example.eventy.common.models.ReservationConfirmationType;
 import org.example.eventy.common.models.SolutionType;
+import org.example.eventy.common.models.Status;
+import org.example.eventy.events.dtos.EventTypeDTO;
+import org.example.eventy.events.models.EventType;
+import org.example.eventy.solutions.dtos.CategoryDTO;
 import org.example.eventy.solutions.dtos.SolutionCardDTO;
+import org.example.eventy.solutions.dtos.services.*;
+import org.example.eventy.solutions.models.Category;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 @Service
 public class ServiceService {
     // @Autowired
     // private ServiceRepository serviceRepository;
+
+    private Collection<org.example.eventy.solutions.models.Service> allServices = new ArrayList<>();
+
+    private void setTestData() {
+        Category category1 = new Category(1L, "Category 1", "Description for Category 1", Status.ACCEPTED);
+        Category category2 = new Category(2L, "Category 2", "Description for Category 2", Status.ACCEPTED);
+
+        List<Category> recommendedCategoriesForEventType1 = new ArrayList<>();
+        recommendedCategoriesForEventType1.add(category1);
+
+        List<Category> recommendedCategoriesForEventType2 = new ArrayList<>();
+        recommendedCategoriesForEventType2.add(category2);
+
+        EventType eventType1 = new EventType(1L, "Event Type 1", "Description for Event Type 1", true, recommendedCategoriesForEventType1);
+        EventType eventType2 = new EventType(2L, "Event Type 2", "Description for Event Type 2", true, recommendedCategoriesForEventType2);
+
+        List<EventType> eventTypeList1 = new ArrayList<>();
+        eventTypeList1.add(eventType1);  // List with first EventType
+
+        List<EventType> eventTypeList2 = new ArrayList<>();
+        eventTypeList2.add(eventType2);
+
+        org.example.eventy.solutions.models.Service service1 = new org.example.eventy.solutions.models.Service(
+                1L, "Service 1", "Description for Service 1", 100.0, 10, new ArrayList<>(Arrays.asList("https://example.com/images/wedding_catering.jpg")),
+                true, true, false, category1, eventTypeList1,
+                null, "Specifics for Service 1", 10, 120, 30, 15, ReservationConfirmationType.AUTOMATIC, null
+        );
+
+        org.example.eventy.solutions.models.Service service2 = new org.example.eventy.solutions.models.Service(
+                2L, "Service 2", "Description for Service 2", 200.0, 20, new ArrayList<>(Arrays.asList("https://example.com/images/wedding_catering.jpg")),
+                true, true, false, category2, eventTypeList2,
+                null, "Specifics for Service 2", 15, 90, 45, 20, ReservationConfirmationType.MANUAL, null
+        );
+
+        org.example.eventy.solutions.models.Service service3 = new org.example.eventy.solutions.models.Service(
+                3L, "Service 3", "Description for Service 3", 150.0, 15, new ArrayList<>(Arrays.asList("https://example.com/images/wedding_catering.jpg")),
+                true, true, false, category1, eventTypeList2,
+                null, "Specifics for Service 3", 20, 100, 60, 30, ReservationConfirmationType.AUTOMATIC, null
+        );
+
+        org.example.eventy.solutions.models.Service service4 = new org.example.eventy.solutions.models.Service(
+                4L, "Service 4", "Description for Service 4", 250.0, 25, new ArrayList<>(Arrays.asList("https://example.com/images/wedding_catering.jpg")),
+                true, true, false, category2, eventTypeList1,
+                null, "Specifics for Service 4", 30, 150, 90, 45, ReservationConfirmationType.MANUAL, null
+        );
+
+        org.example.eventy.solutions.models.Service service5 = new org.example.eventy.solutions.models.Service(
+                5L, "Service 5", "Description for Service 5", 180.0, 18, new ArrayList<>(Arrays.asList("https://example.com/images/wedding_catering.jpg")),
+                true, true, false, category1, eventTypeList2,
+                null, "Specifics for Service 5", 25, 110, 75, 35, ReservationConfirmationType.AUTOMATIC, null
+        );
+
+        allServices.add(service1);
+        allServices.add(service2);
+        allServices.add(service3);
+        allServices.add(service4);
+        allServices.add(service5);
+    }
+
+    public ServiceService() {
+        setTestData();
+    }
+
+    public CreatedServiceDTO createService(CreateServiceDTO createServiceDTO) {
+        CreatedServiceDTO createdServiceDTO = new CreatedServiceDTO();
+        createdServiceDTO.setId(1337L);
+        createdServiceDTO.setName(createServiceDTO.getName());
+        createdServiceDTO.setDescription(createServiceDTO.getDescription());
+        createdServiceDTO.setPrice(createServiceDTO.getPrice());
+        createdServiceDTO.setDiscount(createServiceDTO.getDiscount());
+        createdServiceDTO.setImages(createServiceDTO.getImages());
+        createdServiceDTO.setCategory(createServiceDTO.getCategory());
+        createdServiceDTO.setRelatedEventTypes(createServiceDTO.getRelatedEventTypes());
+        createdServiceDTO.setSpecifics(createServiceDTO.getSpecifics());
+        createdServiceDTO.setMinReservationTime(createServiceDTO.getMinReservationTime());
+        createdServiceDTO.setMaxReservationTime(createServiceDTO.getMaxReservationTime());
+        createdServiceDTO.setReservationDeadline(createServiceDTO.getReservationDeadline());
+        createdServiceDTO.setCancellationDeadline(createServiceDTO.getCancellationDeadline());
+        createdServiceDTO.setAutomaticReservationAcceptance(createServiceDTO.getAutomaticReservationAcceptance());
+        return createdServiceDTO;
+    }
+
+    /*public Collection<GetServiceDTO> getServices(String name, CategoryDTO category, EventTypeDTO eventType, double minPrice, double maxPrice, boolean available) {
+        Collection<GetServiceDTO> getServiceDTOs = new ArrayList<>();
+        allServices.forEach(s -> getServiceDTOs.add(new GetServiceDTO(s)));
+        return getServiceDTOs;
+    }*/
+
+    public Optional<GetServiceDTO> getService(long id) {
+        return Optional.of(new GetServiceDTO(allServices.stream().filter(s -> s.getId() == id).findFirst().get()));
+    }
+
+    public Optional<UpdatedServiceDTO> updateService(UpdateServiceDTO updateServiceDTO) {
+        Optional<org.example.eventy.solutions.models.Service> oldService = allServices.stream().filter(s -> s.getId() == updateServiceDTO.getId()).findFirst();
+        if (!oldService.isPresent()) {
+            return Optional.empty();
+        }
+        UpdatedServiceDTO updatedServiceDTO = new UpdatedServiceDTO();
+        updatedServiceDTO.setId(updateServiceDTO.getId());
+        updatedServiceDTO.setName(updateServiceDTO.getName());
+        updatedServiceDTO.setDescription(updateServiceDTO.getDescription());
+        updatedServiceDTO.setPrice(updateServiceDTO.getPrice());
+        updatedServiceDTO.setDiscount(updateServiceDTO.getDiscount());
+        updatedServiceDTO.setImages(updateServiceDTO.getImages());
+        updatedServiceDTO.setVisible(updateServiceDTO.isVisible());
+        updatedServiceDTO.setAvailable(updateServiceDTO.isAvailable());
+        updatedServiceDTO.setCategory(updateServiceDTO.getCategory());
+        updatedServiceDTO.setRelatedEventTypes(updateServiceDTO.getRelatedEventTypes());
+        updatedServiceDTO.setSpecifics(updateServiceDTO.getSpecifics());
+        updatedServiceDTO.setMinReservationTime(updateServiceDTO.getMinReservationTime());
+        updatedServiceDTO.setMaxReservationTime(updateServiceDTO.getMaxReservationTime());
+        updatedServiceDTO.setReservationDeadline(updateServiceDTO.getReservationDeadline());
+        updatedServiceDTO.setCancellationDeadline(updateServiceDTO.getCancellationDeadline());
+        updatedServiceDTO.setAutomaticReservationAcceptance(updateServiceDTO.getAutomaticReservationAcceptance());
+        return Optional.of(updatedServiceDTO);
+    }
+
+    public void deleteService(long id) {
+        allServices.removeIf(s -> s.getId() == id);
+    }
 
     public ArrayList<SolutionCardDTO> getServices(Pageable pageable) {
         SolutionCardDTO serviceCard1 = new SolutionCardDTO(
@@ -22,7 +148,7 @@ public class ServiceService {
             null, // description - only for Product
             3,
             8,
-            new ArrayList<>(List.of("Wedding", "Reception")),
+            new ArrayList<>(Arrays.asList("Wedding", "Reception")),
             50.00,
             5,
             "https://example.com/images/wedding_catering.jpg",
@@ -40,7 +166,7 @@ public class ServiceService {
             null, // description - only for Product
             2,
             5,
-            new ArrayList<>(List.of("Wedding", "Party")),
+            new ArrayList<>(Arrays.asList("Wedding", "Party")),
             1200.00,
             0,
             "https://example.com/images/music_band.jpg",
@@ -66,7 +192,7 @@ public class ServiceService {
             null, // description - only for Product
             3,
             8,
-            new ArrayList<>(List.of("Wedding", "Reception")),
+            new ArrayList<>(Arrays.asList("Wedding", "Reception")),
             50.00,
             5,
             "https://example.com/images/wedding_catering.jpg",
